MOVED invitation and friendship functionalities to "friendship_module". ADD /friendship_module/ django app to handle friendship related issues.
