+++ Backend/friendship_module/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FriendshipModuleConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'friendship_module'
